refactor(frontend): Add MemberList component
