+++ frontend/src/app/user/member-list/member-list.component.html
@@ -1,0 +1,16 @@
+<ul class="list-group">
+  <li *ngFor="let member of members; let index = index" class="list-group-item d-flex align-items-center">
+    <div class="me-auto">
+      {{ member._user
+      ? member._user.firstName + ' ' + member._user.lastName + ' (' + member._user.username + ')'
+      : 'Loading...' }}
+      <span class="text-muted">&bull; {{ member.userId === owner ? 'Owner' : 'Collaborator' }}</span>
+    </div>
+    <button
+      *ngIf="member.userId !== owner && deleted.observed"
+      type="button" class="btn btn-sm btn-danger bi-person-x"
+      ngbTooltip="Remove Collaborator"
+      (click)="delete(member)"
+    ></button>
+  </li>
+</ul>
